move region and taskcondition classes to their own file, move create regions function to to region class and make it static
